Files structure update of status
